# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LifeLoop API",
    "version": "1.0.0",
    "description": "The backend API powering LifeLoop — scanning, AI recommendations, community, gamification, and notifications. All endpoints except registration/login require a valid session cookie (obtained by logging in)."
  },
  "servers": [{ "url": "/api", "description": "Local server" }],
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "lifeloop_token" }
    },
...
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/community/posts": {
      "get": { "summary": "List/search the community feed", "parameters": [{ "name": "search", "in": "query", "schema": { "type": "string" } }, { "name": "userId", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Posts" } } },
      "post": { "summary": "Create a post (multipart: description, beforeImage, afterImage)", "responses": { "201": { "description": "Post created" } } }
    },
    "/community/posts/{id}": {
      "get": { "summary": "Get a post with its comments", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Post + comments" } } },
      "patch": { "summary": "Edit your own post", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Updated" } } },
      "delete": { "summary": "Delete your own post (or any post, if admin)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Deleted" } } }
    },
    "/community/posts/{id}/like": {
      "post": { "summary": "Like a post", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Liked" } } },
      "delete": { "summary": "Unlike a post", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Unliked" } } }
    },
    "/community/posts/{id}/comments": {
      "post": { "summary": "Comment on a post", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "201": { "description": "Comment created" } } }
    },
    "/community/comments/{id}": {
      "delete": { "summary": "Delete your own comment (or any, if admin)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Deleted" } } }
    },
    "/community/users/{id}/follow": {
      "post": { "summary": "Follow a user", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Followed" } } },
      "delete": { "summary": "Unfollow a user", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Unfollowed" } } }
    },
    "/community/users/{id}": {
      "get": { "summary": "Get a user's public profile (use 'me' for your own)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Profile" } } }
    },
...
    "/community/search": {
      "get": { "summary": "Search posts and people", "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Search results" } } }
    },
...
    "/community/challenges/{id}/join": {
      "post": { "summary": "Join a challenge", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Joined" } } }
    },
...
    "/community/reports": {
      "get": { "summary": "List open reports (admin only)", "responses": { "200": { "description": "Reports" }, "403": { "description": "Admin required" } } }
    },
    "/community/reports/{id}/resolve": {
      "post": { "summary": "Resolve a report: dismiss, remove_content, suspend_user, or ban_user (admin only)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Resolved" }, "403": { "description": "Admin required" } } }
    },
...
  }
}